Bugfix: better FileStorage handling
If a file is uploaded with a WTF.FileField, with no specified
handler (upload_set), it will look for an attribute {field}_handler
where {field} is the name of the field.

If the file was uploaded with a pynuts.UploadField (or a field
inheriting from it), the specified upload_set.save method will
be used for saving the file.

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@kozea.fr>