Default configuration keys defined in Pynuts __init__ method
Default configuration keys (`CSRF_ENABLED`,
PYNUTS_DOCUMENT_REPOSITORY` and `UPLOADS_DEFAULT_DEST`) are
defined in the Pynuts `__init__` method.

As the Pynuts object is to be instanciated after the Flask application,
as the latter is passed as a parameter at init, if the user set
these configuration keys in the Flask config, the user configuration
will be left untouched.

To do son we use the builtin dict.setdefault method.

See also: http://docs.python.org/2/library/stdtypes.html#dict.setdefault

Signed-off-by: Balthazar Rouberol <balthazar.rouberol@kozea.fr>